+++ src/pages/underwriting/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { UnderwritingSystem } from '@/systems/underwriting';
-
-const UnderwritingPage: React.FC = () => {
-  return <UnderwritingSystem />;
-};
-
-export default UnderwritingPage;