+++ src/shared/types/index.ts
@@ -1,6 +1,0 @@
-// Shared Types
-// This file exports all shared types
-
-// Export shared types
-export * from './components';
-// Note: global.d.ts doesn't need to be exported as it's automatically included