+++ src/systems/traffic-light/components/index.ts
@@ -1,9 +1,0 @@
-export { default as CIODashboard } from './CIODashboard';
-export { default as FrontrunSuburbs } from './FrontrunSuburbs';
-export { default as TrafficLightZones } from './TrafficLightZones';
-export { default as FundParameters } from './FundParameters';
-export { default as FundStrategy } from './FundStrategy';
-export { default as MLSystemIntegration } from './MLSystemIntegration';
-export { default as MLEnhancedMap } from './MLEnhancedMap';
-export { default as TrafficLightDashboard } from './TrafficLightDashboard';
-export { default as TrafficLightLayout } from './TrafficLightLayout';