+++ src/systems/traffic-light/utils/index.ts
@@ -1,4 +1,0 @@
-// Traffic Light System Utilities
-// This file will export all utilities for the Traffic Light System
-
-// Utilities will be added as they are migrated