+++ src/systems/underwriting/modules/marketing/store/index.ts
@@ -1,9 +1,0 @@
-// Sales & Marketing Dashboard Module Store
-// This file exports all stores for the Sales & Marketing Dashboard module
-
-// Export stores as they are implemented
-export * from './targetPropertyStore';
-export * from './campaignStore';
-export * from './leadStore';
-export * from './mailchimpStore';
-export * from './filterStore';