+++ src/systems/underwriting/modules/pipeline/store/index.ts
@@ -1,8 +1,0 @@
-// Pipeline Management Module Store
-// This file exports all stores for the Pipeline Management module
-
-// Export stores as they are implemented
-export * from './dealStore';
-export * from './rankingStore';
-export * from './taskStore';
-export * from './filterStore';