+++ src/systems/underwriting/modules/underwriting/store/index.ts
@@ -1,9 +1,0 @@
-// Underwriting & Origination Module Store
-// This file exports all stores for the Underwriting & Origination module
-
-// Export stores as they are implemented
-export * from './applicationStore';
-export * from './propertyStore';
-export * from './riskStore';
-export * from './decisionStore';
-export * from './parameterStore';