+++ src/utils/chartConfig.ts
@@ -1,6 +1,0 @@
-import { Chart, registerables } from 'chart.js';
-
-// Register all Chart.js components
-Chart.register(...registerables);
-
-export default Chart;